Refactor repositories and entities

+++ persistence/src/main/java/com/bopera/pointofsales/repository/RoomRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface RoomRepository extends BaseRepository<Room, Integer> {
+public interface RoomRepository extends BaseRepository<Room, Long> {
     Optional<Room> findTopByOrderBySortingDesc();
 
     List<Room> findAllByOrderBySortingDesc();
